refactor: fileGenerator 새로운 파일 이름만 생성하도록 변경
